Combine kinds.js and index.js
