+++ core/textify/proto.js
@@ -1,18 +1,16 @@
 import rewrite from "../../pack/rewrite.js";
 
 import { check as recheck } from "./util/check.js";
-import { index as reindex } from "./core/index.js";
+import { kinds as reindex } from "./core/kinds.js";
 import { store as restore } from "./core/store.js";
 import { group as regroup } from "./core/group.js";
-import { kinds as kindsof } from "./core/kinds.js";
 
 const prototype = {
 
   recheck,
   reindex,
   restore,
-  regroup,
-  kindsof
+  regroup
 
 }
